refactor(lsp): consolidate identifier-at-position, rangesEqual, and comment scanning into documentUtils
