+++ src/languageServer/definition.ts
@@ -5,6 +5,7 @@ import { Location, Position, Range } from "vscode-languageserver/node";
 import { TextDocument } from "vscode-languageserver-textdocument";
 import { URI } from "vscode-uri";
 
+import { getIdentifierAtPosition, getPrecedingQualifiedName, rangesEqual } from "./documentUtils";
 import { getBlockEndLine, getDocumentLines, getDocumentStructure } from "./symbols";
 
 export type DefinitionEntry = {
@@ -83,13 +84,14 @@ export function getMatchingDefinition(document: TextDocument, symbol: SymbolAtPo
 
 export function getDefinitionEntries(document: TextDocument): DefinitionEntry[] {
     const lines = getDocumentLines(document);
-    const entries: DefinitionEntry[] = getDocumentStructure(document).map((item) => ({
+    const structure = getDocumentStructure(document);
+    const entries: DefinitionEntry[] = structure.map((item) => ({
         detail: item.detail,
         name: item.name,
         range: Range.create(item.line, item.nameStart, item.line, item.nameEnd),
     }));
 
-    for (const item of getDocumentStructure(document)) {
+    for (const item of structure) {
         if (item.detail !== "enum") {
             continue;
         }
@@ -116,58 +118,17 @@ export function getDefinitionEntries(document: TextDocument): DefinitionEntry[] 
 }
 
 export function getSymbolAtPosition(document: TextDocument, position: Position): SymbolAtPosition | null {
-    const text = document.getText();
-    let offset = document.offsetAt(position);
-
-    if (!isWordCharacter(text.charAt(offset)) && offset > 0 && isWordCharacter(text.charAt(offset - 1))) {
-        offset--;
-    }
-
-    if (!isWordCharacter(text.charAt(offset))) {
+    const positioned = getIdentifierAtPosition(document, position, /[A-Za-z0-9_]/, /^[A-Za-z0-9_]+$/);
+    if (!positioned) {
         return null;
     }
 
-    const startOffset = findWordStart(text, offset);
-    const endOffset = findWordEnd(text, offset);
-    const name = text.slice(startOffset, endOffset);
-    const precedingName = getPrecedingQualifiedName(text, startOffset);
-
+    const startOffset = document.offsetAt(positioned.range.start);
     return {
-        name,
-        precedingName,
-        range: Range.create(document.positionAt(startOffset), document.positionAt(endOffset)),
+        name: positioned.text,
+        precedingName: getPrecedingQualifiedName(document.getText(), startOffset),
+        range: positioned.range,
     };
-}
-
-function getPrecedingQualifiedName(text: string, wordStartOffset: number): string | undefined {
-    if (wordStartOffset < 2 || text.charAt(wordStartOffset - 1) !== ".") {
-        return undefined;
-    }
-
-    const previousWordEnd = wordStartOffset - 1;
-    const previousWordStart = findWordStart(text, previousWordEnd - 1);
-    const previousWord = text.slice(previousWordStart, previousWordEnd);
-    return previousWord.length > 0 ? previousWord : undefined;
-}
-
-function findWordStart(text: string, offset: number): number {
-    let startOffset = offset;
-    while (startOffset > 0 && isWordCharacter(text.charAt(startOffset - 1))) {
-        startOffset--;
-    }
-    return startOffset;
-}
-
-function findWordEnd(text: string, offset: number): number {
-    let endOffset = offset + 1;
-    while (endOffset < text.length && isWordCharacter(text.charAt(endOffset))) {
-        endOffset++;
-    }
-    return endOffset;
-}
-
-function isWordCharacter(character: string): boolean {
-    return /[A-Za-z0-9_]/.test(character);
 }
 
 export async function getWorkspaceDocuments(
@@ -207,15 +168,6 @@ export async function getWorkspaceDocuments(
     }
 
     return [...workspaceDocuments.values()];
-}
-
-function rangesEqual(left: Range, right: Range): boolean {
-    return (
-        left.start.line === right.start.line &&
-        left.start.character === right.start.character &&
-        left.end.line === right.end.line &&
-        left.end.character === right.end.character
-    );
 }
 
 async function findOpyFiles(root: string): Promise<string[]> {
